Move `getTranslatedEntityName` to the common utils
