+++ frontend/src/metabase/common/utils/model-names.ts
