--- frontend/src/metabase/nav/utils/index.js
@@ -1,1 +1,0 @@
-export * from "./model-names";
