+++ frontend/src/metabase/search/components/InfoText.tsx
@@ -9,7 +9,7 @@ import Schema from "metabase/entities/schemas";
 import Database from "metabase/entities/databases";
 import Table from "metabase/entities/tables";
 import { PLUGIN_COLLECTIONS } from "metabase/plugins";
-import { getTranslatedEntityName } from "metabase/nav/utils";
+import { getTranslatedEntityName } from "metabase/common/utils/model-names";
 
 import type { Collection } from "metabase-types/api";
 import type { WrappedResult } from "metabase/search/types";
